# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto
  description: |-
    WasaPhoto is a social network where users can upload their pictures and follow friends to see their pictures and interact with them.
  version: 1.0.0
servers: 
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Spazio-D/WasaPhoto/1.0.0
  - url: "http://localhost:3000"
security:
  - bearerAuth: [] 
...
paths:
...
  /users/{user_id}/posts:
    parameters:
      - name: user_id
        in: path
        description: The user id
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    post:
      tags: ["Post"]
      description: Upload a photo and create a post.
      summary: Upload a photo and create a post.
      operationId: uploadPhoto
      requestBody:
        content:
          multipart/form-data:
            schema:
              description: The image to upload.
              type: object
              properties:
                image: { $ref: "#/components/schemas/Image" }
      responses:
        "201":
          description: Post created successfully
          content:
            text/plain:
              schema:
                description: Request success
                type: string
                example: "Resource successfully created"
            multipart/form-data:
              schema:
                description: The post created
                type: object
                properties:
                  post: { $ref: "#/components/schemas/Post" }
                  time: { $ref: "#/components/schemas/TimeStamp" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    get:
      tags: ["Post"]
      description: Get posts of the user
      summary: Get posts
      operationId: getPosts
      responses:
        "200":
          description: Posts got successfully
          content:
            text/plain:
              schema:
                description: Request success
                type: string
                example: "Action successfully completed"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  #FOLLOW USER
  #UNFOLLOW USER
...
  /users/{user_id}/posts/{post_id}/likes/{liker_id}:
    parameters: 
      - name: user_id
        in: path
        description: The user id of the post owner
        schema: { $ref: "#/components/schemas/UserID" }
        required: true
      - name: post_id
        in: path
        description: The post id
        schema: { $ref: "#/components/schemas/PostID" }
        required: true
      - name: liker_id
        in: path
        description: The user id of the user who wants to like or unlike the post
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    put:
      tags: ["Post"]
      summary: Like the post
      operationId: likePhoto
      description: Like a post
      responses:
        "201":
          description: Like added successfully
          content:
            text/plain:
              schema:
                description: Request success
                type: string
                example: "Resource successfully created"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

    delete:
      tags: ["Post"]
      summary: Unlike a post
      operationId: unlikePhoto
      description: Unlike the post
      responses:
        "200":
          description: Like removed successfully
          content:
            text/plain:
              schema: 
                description: Request success
                type: string
                example: "Action successfully completed"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }

  #COMMENT POST
...
components:
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
  schemas:
...
    UserID:
      title: UserID
      description: Unique numeric id of the user
      type: integer
      example: 348
      minimum: 0
      maximum: 9999999999999
      readOnly: true
...
    TimeStamp:
      title: TimeStamp
      description: Describe a time in the YY-MM-DD HH:MM:SS format
      type: string
      format: date-time
      pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$
      example: "2023-11-12T14:59:37Z"
      minLength: 20
      maxLength: 20

    #POST AND IMAGE  
    PostID:
      title: PostID
      description: Unique numeric id of the post
      type: integer
      example: 41
      minimum: 0
      maximum: 99999999
      readOnly: true
    Image:
      title: Image
      description: A image represented in base64
      type: string
      format: byte
      minLength: 0
      maxLength: 1000000000  
    Post:
      title: Post
      description: A pubblication of a picture of a user on his own profile
      type: object
      properties:
        image: { $ref: "#/components/schemas/Image" }
        post_id: { $ref: "#/components/schemas/PostID" }
        likes_count:
          description: Number of likes of the post
          type: integer
          example: 69
          minimum: 0
          maximum: 99999999999
        comments_count: 
          description: Number of comments of the post
          type: integer
          example: 666
          minimum: 0
          maximum: 99999999999
        like_check:
          type: boolean
          description: True if the post is liked by the logged user
          example: true
        timestamp: { $ref: "#/components/schemas/TimeStamp" }
...
  responses:
...
    BadRequest:
      description: |-
        The server cannot or will not process the request due to an apparent client error (e.g., malformed request syntax, size too large, invalid request message framing, or deceptiverequest routing).
      content:
        text/plain:
          schema:
            description: Bad Request
            type: string
            example: "Bad Request, maybe sintax error"
            minLength: 5
            maxLength: 100

    #401
    Unauthorized:
      description: |-
            Used when authentication is required and has failed or has not yet been provided. Semantically means "unauthorised", the user does not have valid authentication credentials for the target resource.
      content:
        text/plain:
          schema:
            description: Unauthorized
            type: string
            example: "Unauthorized"
            minLength: 5
            maxLength: 100

    #403
    Forbidden:
      description: |-
        The request contained valid data and was understood by the server, but the server is refusing action. This may be due to the user not having the necessary permissions for a resource or needing an account of some sort, or attempting a prohibited action (e.g. creating a duplicate record where only one is allowed).
      content:
        text/plain:
          schema:
            description: Forbidden
            type: string
            example: "Forbidden"
            minLength: 5
            maxLength: 100

    #500
    InternalServerError:
      description: |-
        A generic error message, given when an unexpected condition was encountered and no more specific message is suitable.
      content:
        text/plain:
          schema:
            description: Internal Server Error
            type: string
            example: "Internal Server Error"
            minLength: 5
            maxLength: 100

    #503
    ServiceUnavailable:
      description: |-
        The server cannot handle the request (because it is overloaded or down for maintenance). Generally, this is a temporary state.      
      content:
        text/plain:
          schema:
            description: Service Unavailable
            type: string
            example: "Service Unavailable"
            minLength: 5
            maxLength: 100